requirements.txt
APScheduler==3.10.4
attrs==23.2.0
black==24.3.0
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-BasicAuth==0.2.0
Flask-HTTPAuth==4.8.0
Flask-json-schema==0.0.5
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
Xmlify==0.1.1
